finalized new queue

+++ src/erc20/interfaces/IERC20Pool.sol
@@ -301,7 +301,7 @@ interface IERC20Pool is IScaledPool {
         external
         view
         returns (
-            uint256 kickTime_,
+            uint128 kickTime_,
             uint256 referencePrice_,
             uint256 bondFactor_,
             uint256 bondSize_
